pyproject.toml
[project]
name = "wyu-navigator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.125.0",
    "loguru>=0.7.3",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
